+++ kernel/auditsc.c
@@ -2952,7 +2952,7 @@ void audit_seccomp(unsigned long syscall, long signr, int code)
 	audit_log_task(ab);
 	audit_log_format(ab, " sig=%ld arch=%x syscall=%ld compat=%d ip=0x%lx code=0x%x",
 			 signr, syscall_get_arch(current), syscall,
-			 in_compat_syscall(), KSTK_EIP(current), code);
+			 in_compat_syscall(), (unsigned long)KSTK_EIP(current), code);
 	audit_log_end(ab);
 }
